# New approach using pyarrow for IO