--- a/cellxgene_schema_cli/requirements.txt
+++ b/cellxgene_schema_cli/requirements.txt
@@ -6,6 +6,7 @@ dask[array, dataframe]==2024.12.0
 filelock<4
 numpy<3
 pandas>2,<3
+pyarrow>=19
 PyYAML<7
 scipy<1.15 # broken in 1.15.0 see https://github.com/chanzuckerberg/single-cell-curation/issues/1165
 semver<4