## Reason for Change

Speed up conversion to parquet by skipping intermediate decompressed tsp

## Changes

- This uses the iterative CSV reader and iterative parquet writer from pyarrow.

## Testing

- Either list QA steps or reasoning you feel QA is unnecessary
- Reminder For CLI changes: upon merge, contact Lattice for final sign-off. Do not release a new cellxgene-schema 
version to PyPI without explicit QA + sign-off from Lattice on all functional CLI changes. They may install the package
version at HEAD of main with 
```
pip install git+https://github.com/chanzuckerberg/single-cell-curation/@main#subdirectory=cellxgene_schema_cli
```

## Notes for Reviewer

[Here is a small notebook](https://gist.github.com/ivirshup/248a2931d6e78cf72dedd9520af0e8de) where I took some performance metrics. I'm seeing about a 3x speed up on conversion and 1/5th memory usage

I've also timed the test suite on my branch against main, where we also see some improvements:

#### this branch

```
5.53s call     cellxgene_schema_cli/tests/test_atac_seq.py::TestProcessFragment::test_write_algorithms[pysam]
5.27s call     cellxgene_schema_cli/tests/test_atac_seq.py::TestProcessFragment::test_source_file_compression[fragments.tsv.bgz]
5.26s call     cellxgene_schema_cli/tests/test_atac_seq.py::TestProcessFragment::test_source_file_compression[fragments_sorted.tsv.gz]
5.12s call     cellxgene_schema_cli/tests/test_atac_seq.py::TestProcessFragment::test_write_algorithms[None]
5.12s call     cellxgene_schema_cli/tests/test_atac_seq.py::TestProcessFragment::test_write_algorithms[cli]
```

#### main

```
7.40s call     cellxgene_schema_cli/tests/test_atac_seq.py::TestProcessFragment::test_write_algorithms[pysam]
7.07s call     cellxgene_schema_cli/tests/test_atac_seq.py::TestProcessFragment::test_source_file_compression[fragments_sorted.tsv.gz]
7.07s call     cellxgene_schema_cli/tests/test_atac_seq.py::TestProcessFragment::test_write_algorithms[None]
7.06s call     cellxgene_schema_cli/tests/test_atac_seq.py::TestProcessFragment::test_source_file_compression[fragments.tsv.bgz]
7.02s call     cellxgene_schema_cli/tests/test_atac_seq.py::TestProcessFragment::test_write_algorithms[cli]
```